requirements.txt
streamlit
groq
sentence-transformers
torch
chromadb
rank-bm25
pypdf
python-dotenv
scikit-learn
numpy